Added several changes in step of transpilation code
