+++ packages/GlassComponent/.babelrc
@@ -1,0 +1,15 @@
+{
+    "presets": [
+        "@babel/preset-env",
+        "@babel/preset-react",
+        "@babel/preset-typescript"
+    ],
+    "plugins": [
+        [
+            "babel-plugin-styled-components",
+            {
+                "ssr": false
+            }
+        ]
+    ]
+}
