--- packages/GlassComponent/glass/src/CreditCard/styles.ts
@@ -1,88 +1,0 @@
-import styled from "styled-components"
-
-export const Section = styled.div`
-    width: 100%;
-    height: 100%;
-
-    display: flex;
-    flex-flow: column;
-    justify-content: space-between;
-`
-
-export const CardHeader = styled.div`
-    width: 100%;
-    height: 20px;
-
-    display: flex;
-    flex-flow: row wrap;
-`
-
-export const TitleCard = styled.p`
-    font-family: 'arial, consolas,"Roboto"', sans-serif;
-    font-weight: 700;
-    font-size: 1.0em;
-    color: #FFFF;
-`
-
-export const CardChip = styled.div`
-    width: 30px;
-    height: 23px;
-
-    & img {
-        max-width: 100%;
-    }
-`
-
-export const CardNumber = styled.div`
-    width: 100%;
-    height: 15px;
-    /* border:1px solid blue; */
-
-    span {
-        font-family: "'Lato Hairline', Lato",sans-serif;
-        font-weight: 500;
-        font-size: 0.7em;
-        color: #FFF;
-        letter-spacing: 3px;
-    }
-
-    display: flex;
-    align-items: center;
-`
-
-export const CardName = styled.div`
-    width: 100%;
-    height: 15px;
-    /* border:1px solid blue; */
-
-    span {
-        font-family: "'Lato Hairline', Lato",sans-serif;
-        font-weight: lighter;
-        font-size: 0.7em;
-        color: #FFF;
-        letter-spacing: 0.7px;
-    }
-
-    display: flex;
-    align-items: center;
-`
-
-export const CardDateValid = styled.div`
-    width: 100%;
-    height: 100%;
-    /* border:1px solid blue; */
-
-    span {
-        font-family: "'Lato Hairline', Lato",sans-serif;
-        font-weight: lighter;
-        font-size: 0.5em;
-        color: #FFF;
-        letter-spacing: 0.7px;
-    }
-
-
-    display: flex;
-    justify-content: flex-end;
-    align-items: flex-end;
-
-`
