--- packages/GlassComponent/glass/src/styles/global.css
@@ -1,14 +1,0 @@
-*,*::before, *::after {
-    outline: none;
-    padding: 0px;
-    margin: 0px;
-    border: none;
-    box-sizing: border-box;
-}
-
-html,body, #root {
-    height: 100vw;
-    font-family: "Roboto", sans-serif;
-    font-size:17px;
-    background: linear-gradient(90deg,#55439E,#952FFF)
-}
