+++ packages/GlassComponent/index.ts
@@ -1,0 +1,9 @@
+import inputGlass from "./src/InputGlass";
+import creditCardGlass from "./src/CreditCardGlass";
+import boxMessageGlass from "./src/BoxMessageGlass";
+import boxChatGlass from "./src/BoxChatGlass";
+
+export const InputGlass = inputGlass
+export const CreditCardGlass = creditCardGlass
+export const BoxMessageGlass = boxMessageGlass
+export const BoxChatGlass = boxChatGlass
