+++ packages/GlassComponent/package.json
@@ -1,8 +1,8 @@
 {
     "name": "glasscomponentss",
-    "version": "1.0.5",
+    "version": "1.1.3",
     "description": "A package with styled components in glassmorphism",
-    "main": "index.ts",
+    "main": "index.js",
     "keywords": [
         "npm",
         "glass",
@@ -12,9 +12,15 @@
         "styleds"
     ],
     "author": "ONEDEV <Pedro Guilherme>",
-    "license": "MIT",
+    "license": "CC",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/PedroGuilhermeFariaDuarte/ONEDEV/tree/main/packages/GlassComponent"
+    },
     "scripts": {
-        "glass:test": "react-scripts test"
+        "glass:test": "react-scripts test",
+        "glass:build": "webpack",
+        "glass:tsc": "tsc ./src/index.ts"
     },
     "dependencies": {
         "@testing-library/jest-dom": "^5.11.9",
@@ -32,6 +38,19 @@
         "web-vitals": "^1.1.0"
     },
     "devDependencies": {
-        "@types/styled-components": "^5.1.7"
+        "@babel/preset-env": "^7.12.13",
+        "@babel/preset-react": "^7.12.13",
+        "@babel/preset-typescript": "^7.12.13",
+        "@types/styled-components": "^5.1.7",
+        "awesome-typescript-loader": "^5.2.1",
+        "babel-core": "^6.26.3",
+        "babel-loader": "8.1.0",
+        "babel-plugin-styled-components": "^1.12.0",
+        "babel-preset-env": "^1.7.0",
+        "babel-preset-react": "^6.24.1",
+        "babel-preset-stage-0": "^6.24.1",
+        "ts-loader": "^8.0.15",
+        "webpack": "^5.21.1",
+        "webpack-cli": "^4.5.0"
     }
 }
