+++ packages/GlassComponent/src/BoxChatGlass/index.js
@@ -1,0 +1,19 @@
+/* eslint-disable new-parens */
+/* eslint-disable strict */
+/* eslint-disable no-new-wrappers */
+/* eslint-disable no-func-assign */
+/* eslint-disable no-useless-escape */
+/* eslint-disable eqeqeq */
+/* eslint-disable no-loop-func */
+/* eslint-disable no-use-before-define */
+/* eslint-disable no-fallthrough */
+/* eslint-disable no-labels */
+/* eslint-disable default-case */
+/* eslint-disable no-sequences */
+/* eslint-disable no-mixed-operators */
+/* eslint-disable no-label-var */
+/* eslint-disable no-unused-expressions */
+/* eslint-disable  no-undef */
+
+/*! For license information please see index.js.LICENSE.txt */
+module.exports = (() => { var e = { 646: (e, t, r) => { "use strict"; r.r(t), r.d(t, { default: () => Be }); const n = require("react"); var a = r.n(n), o = r(864), i = r(774), s = r.n(i); const c = function (e) { function t(e, n, c, l, f) { for (var p, h, g, m, S, x = 0, C = 0, k = 0, A = 0, P = 0, T = 0, j = g = p = 0, B = 0, N = 0, z = 0, L = 0, F = c.length, D = F - 1, H = "", G = "", W = "", Y = ""; B < F;) { if (h = c.charCodeAt(B), B === D && 0 !== C + A + k + x && (0 !== C && (h = 47 === C ? 10 : 47), A = k = x = 0, F++, D++), 0 === C + A + k + x) { if (B === D && (0 < N && (H = H.replace(d, "")), 0 < H.trim().length)) { switch (h) { case 32: case 9: case 59: case 13: case 10: break; default: H += c.charAt(B) }h = 59 } switch (h) { case 123: for (p = (H = H.trim()).charCodeAt(0), g = 1, L = ++B; B < F;) { switch (h = c.charCodeAt(B)) { case 123: g++; break; case 125: g--; break; case 47: switch (h = c.charCodeAt(B + 1)) { case 42: case 47: e: { for (j = B + 1; j < D; ++j)switch (c.charCodeAt(j)) { case 47: if (42 === h && 42 === c.charCodeAt(j - 1) && B + 2 !== j) { B = j + 1; break e } break; case 10: if (47 === h) { B = j + 1; break e } }B = j } }break; case 91: h++; case 40: h++; case 34: case 39: for (; B++ < D && c.charCodeAt(B) !== h;); }if (0 === g) break; B++ } switch (g = c.substring(L, B), 0 === p && (p = (H = H.replace(u, "").trim()).charCodeAt(0)), p) { case 64: switch (0 < N && (H = H.replace(d, "")), h = H.charCodeAt(1)) { case 100: case 109: case 115: case 45: N = n; break; default: N = $ }if (L = (g = t(n, N, g, h, f + 1)).length, 0 < _ && (S = s(3, g, N = r($, H, z), n, I, O, L, h, f, l), H = N.join(""), void 0 !== S && 0 === (L = (g = S.trim()).length) && (h = 0, g = "")), 0 < L) switch (h) { case 115: H = H.replace(w, i); case 100: case 109: case 45: g = H + "{" + g + "}"; break; case 107: g = (H = H.replace(y, "$1 $2")) + "{" + g + "}", g = 1 === E || 2 === E && o("@" + g, 3) ? "@-webkit-" + g + "@" + g : "@" + g; break; default: g = H + g, 112 === l && (G += g, g = "") } else g = ""; break; default: g = t(n, r(n, H, z), g, l, f + 1) }W += g, g = z = N = j = p = 0, H = "", h = c.charCodeAt(++B); break; case 125: case 59: if (1 < (L = (H = (0 < N ? H.replace(d, "") : H).trim()).length)) switch (0 === j && (p = H.charCodeAt(0), 45 === p || 96 < p && 123 > p) && (L = (H = H.replace(" ", ":")).length), 0 < _ && void 0 !== (S = s(1, H, n, e, I, O, G.length, l, f, l)) && 0 === (L = (H = S.trim()).length) && (H = "\0\0"), p = H.charCodeAt(0), h = H.charCodeAt(1), p) { case 0: break; case 64: if (105 === h || 99 === h) { Y += H + c.charAt(B); break } default: 58 !== H.charCodeAt(L - 1) && (G += a(H, p, h, H.charCodeAt(2))) }z = N = j = p = 0, H = "", h = c.charCodeAt(++B) } } switch (h) { case 13: case 10: 47 === C ? C = 0 : 0 === 1 + p && 107 !== l && 0 < H.length && (N = 1, H += "\0"), 0 < _ * M && s(0, H, n, e, I, O, G.length, l, f, l), O = 1, I++; break; case 59: case 125: if (0 === C + A + k + x) { O++; break } default: switch (O++, m = c.charAt(B), h) { case 9: case 32: if (0 === A + x + C) switch (P) { case 44: case 58: case 9: case 32: m = ""; break; default: 32 !== h && (m = " ") }break; case 0: m = "\\0"; break; case 12: m = "\\f"; break; case 11: m = "\\v"; break; case 38: 0 === A + C + x && (N = z = 1, m = "\f" + m); break; case 108: if (0 === A + C + x + R && 0 < j) switch (B - j) { case 2: 112 === P && 58 === c.charCodeAt(B - 3) && (R = P); case 8: 111 === T && (R = T) }break; case 58: 0 === A + C + x && (j = B); break; case 44: 0 === C + k + A + x && (N = 1, m += "\r"); break; case 34: case 39: 0 === C && (A = A === h ? 0 : 0 === A ? h : A); break; case 91: 0 === A + C + k && x++; break; case 93: 0 === A + C + k && x--; break; case 41: 0 === A + C + x && k--; break; case 40: if (0 === A + C + x) { if (0 === p) switch (2 * P + 3 * T) { case 533: break; default: p = 1 }k++ } break; case 64: 0 === C + k + A + x + j + g && (g = 1); break; case 42: case 47: if (!(0 < A + x + k)) switch (C) { case 0: switch (2 * h + 3 * c.charCodeAt(B + 1)) { case 235: C = 47; break; case 220: L = B, C = 42 }break; case 42: 47 === h && 42 === P && L + 2 !== B && (33 === c.charCodeAt(L + 2) && (G += c.substring(L, B + 1)), m = "", C = 0) } }0 === C && (H += m) }T = P, P = h, B++ } if (0 < (L = G.length)) { if (N = n, 0 < _ && void 0 !== (S = s(2, G, N, e, I, O, L, l, f, l)) && 0 === (G = S).length) return Y + G + W; if (G = N.join(",") + "{" + G + "}", 0 != E * R) { switch (2 !== E || o(G, 2) || (R = 0), R) { case 111: G = G.replace(v, ":-moz-$1") + G; break; case 112: G = G.replace(b, "::-webkit-input-$1") + G.replace(b, "::-moz-$1") + G.replace(b, ":-ms-input-$1") + G }R = 0 } } return Y + G + W } function r(e, t, r) { var a = t.trim().split(g); t = a; var o = a.length, i = e.length; switch (i) { case 0: case 1: var s = 0; for (e = 0 === i ? "" : e[0] + " "; s < o; ++s)t[s] = n(e, t[s], r).trim(); break; default: var c = s = 0; for (t = []; s < o; ++s)for (var l = 0; l < i; ++l)t[c++] = n(e[l] + " ", a[s], r).trim() }return t } function n(e, t, r) { var n = t.charCodeAt(0); switch (33 > n && (n = (t = t.trim()).charCodeAt(0)), n) { case 38: return t.replace(m, "$1" + e.trim()); case 58: return e.trim() + t.replace(m, "$1" + e.trim()); default: if (0 < 1 * r && 0 < t.indexOf("\f")) return t.replace(m, (58 === e.charCodeAt(0) ? "" : "$1") + e.trim()) }return e + t } function a(e, t, r, n) { var i = e + ";", s = 2 * t + 3 * r + 4 * n; if (944 === s) { e = i.indexOf(":", 9) + 1; var c = i.substring(e, i.length - 1).trim(); return c = i.substring(0, e).trim() + c + ";", 1 === E || 2 === E && o(c, 1) ? "-webkit-" + c + c : c } if (0 === E || 2 === E && !o(i, 1)) return i; switch (s) { case 1015: return 97 === i.charCodeAt(10) ? "-webkit-" + i + i : i; case 951: return 116 === i.charCodeAt(3) ? "-webkit-" + i + i : i; case 963: return 110 === i.charCodeAt(5) ? "-webkit-" + i + i : i; case 1009: if (100 !== i.charCodeAt(4)) break; case 969: case 942: return "-webkit-" + i + i; case 978: return "-webkit-" + i + "-moz-" + i + i; case 1019: case 983: return "-webkit-" + i + "-moz-" + i + "-ms-" + i + i; case 883: if (45 === i.charCodeAt(8)) return "-webkit-" + i + i; if (0 < i.indexOf("image-set(", 11)) return i.replace(P, "$1-webkit-$2") + i; break; case 932: if (45 === i.charCodeAt(4)) switch (i.charCodeAt(5)) { case 103: return "-webkit-box-" + i.replace("-grow", "") + "-webkit-" + i + "-ms-" + i.replace("grow", "positive") + i; case 115: return "-webkit-" + i + "-ms-" + i.replace("shrink", "negative") + i; case 98: return "-webkit-" + i + "-ms-" + i.replace("basis", "preferred-size") + i }return "-webkit-" + i + "-ms-" + i + i; case 964: return "-webkit-" + i + "-ms-flex-" + i + i; case 1023: if (99 !== i.charCodeAt(8)) break; return "-webkit-box-pack" + (c = i.substring(i.indexOf(":", 15)).replace("flex-", "").replace("space-between", "justify")) + "-webkit-" + i + "-ms-flex-pack" + c + i; case 1005: return p.test(i) ? i.replace(f, ":-webkit-") + i.replace(f, ":-moz-") + i : i; case 1e3: switch (t = (c = i.substring(13).trim()).indexOf("-") + 1, c.charCodeAt(0) + c.charCodeAt(t)) { case 226: c = i.replace(S, "tb"); break; case 232: c = i.replace(S, "tb-rl"); break; case 220: c = i.replace(S, "lr"); break; default: return i }return "-webkit-" + i + "-ms-" + c + i; case 1017: if (-1 === i.indexOf("sticky", 9)) break; case 975: switch (t = (i = e).length - 10, s = (c = (33 === i.charCodeAt(t) ? i.substring(0, t) : i).substring(e.indexOf(":", 7) + 1).trim()).charCodeAt(0) + (0 | c.charCodeAt(7))) { case 203: if (111 > c.charCodeAt(8)) break; case 115: i = i.replace(c, "-webkit-" + c) + ";" + i; break; case 207: case 102: i = i.replace(c, "-webkit-" + (102 < s ? "inline-" : "") + "box") + ";" + i.replace(c, "-webkit-" + c) + ";" + i.replace(c, "-ms-" + c + "box") + ";" + i }return i + ";"; case 938: if (45 === i.charCodeAt(5)) switch (i.charCodeAt(6)) { case 105: return c = i.replace("-items", ""), "-webkit-" + i + "-webkit-box-" + c + "-ms-flex-" + c + i; case 115: return "-webkit-" + i + "-ms-flex-item-" + i.replace(C, "") + i; default: return "-webkit-" + i + "-ms-flex-line-pack" + i.replace("align-content", "").replace(C, "") + i }break; case 973: case 989: if (45 !== i.charCodeAt(3) || 122 === i.charCodeAt(4)) break; case 931: case 953: if (!0 === A.test(e)) return 115 === (c = e.substring(e.indexOf(":") + 1)).charCodeAt(0) ? a(e.replace("stretch", "fill-available"), t, r, n).replace(":fill-available", ":stretch") : i.replace(c, "-webkit-" + c) + i.replace(c, "-moz-" + c.replace("fill-", "")) + i; break; case 962: if (i = "-webkit-" + i + (102 === i.charCodeAt(5) ? "-ms-" + i : "") + i, 211 === r + n && 105 === i.charCodeAt(13) && 0 < i.indexOf("transform", 10)) return i.substring(0, i.indexOf(";", 27) + 1).replace(h, "$1-webkit-$2") + i }return i } function o(e, t) { var r = e.indexOf(1 === t ? ":" : "{"), n = e.substring(0, 3 !== t ? r : 10); return r = e.substring(r + 1, e.length - 1), j(2 !== t ? n : n.replace(k, "$1"), r, t) } function i(e, t) { var r = a(t, t.charCodeAt(0), t.charCodeAt(1), t.charCodeAt(2)); return r !== t + ";" ? r.replace(x, " or ($1)").substring(4) : "(" + t + ")" } function s(e, t, r, n, a, o, i, s, c, u) { for (var d, f = 0, p = t; f < _; ++f)switch (d = T[f].call(l, e, p, r, n, a, o, i, s, c, u)) { case void 0: case !1: case !0: case null: break; default: p = d }if (p !== t) return p } function c(e) { return void 0 !== (e = e.prefix) && (j = null, e ? "function" != typeof e ? E = 1 : (E = 2, j = e) : E = 0), c } function l(e, r) { var n = e; if (33 > n.charCodeAt(0) && (n = n.trim()), n = [n], 0 < _) { var a = s(-1, r, n, n, I, O, 0, 0, 0, 0); void 0 !== a && "string" == typeof a && (r = a) } var o = t($, n, r, 0, 0); return 0 < _ && void 0 !== (a = s(-2, o, n, n, I, O, o.length, 0, 0, 0)) && (o = a), R = 0, O = I = 1, o } var u = /^\0+/g, d = /[\0\r\f]/g, f = /: */g, p = /zoo|gra/, h = /([,: ])(transform)/g, g = /,\r+?/g, m = /([\t\r\n ])*\f?&/g, y = /@(k\w+)\s*(\S*)\s*/, b = /::(place)/g, v = /:(read-only)/g, S = /[svh]\w+-[tblr]{2}/, w = /\(\s*(.*)\s*\)/g, x = /([\s\S]*?);/g, C = /-self|flex-/g, k = /[^]*?(:[rp][el]a[\w-]+)[^]*/, A = /stretch|:\s*\w+\-(?:conte|avail)/, P = /([^-])(image-set\()/, O = 1, I = 1, R = 0, E = 1, $ = [], T = [], _ = 0, j = null, M = 0; return l.use = function e(t) { switch (t) { case void 0: case null: _ = T.length = 0; break; default: if ("function" == typeof t) T[_++] = t; else if ("object" == typeof t) for (var r = 0, n = t.length; r < n; ++r)e(t[r]); else M = 0 | !!t }return e }, l.set = c, void 0 !== e && c(e), l }, l = { animationIterationCount: 1, borderImageOutset: 1, borderImageSlice: 1, borderImageWidth: 1, boxFlex: 1, boxFlexGroup: 1, boxOrdinalGroup: 1, columnCount: 1, columns: 1, flex: 1, flexGrow: 1, flexPositive: 1, flexShrink: 1, flexNegative: 1, flexOrder: 1, gridRow: 1, gridRowEnd: 1, gridRowSpan: 1, gridRowStart: 1, gridColumn: 1, gridColumnEnd: 1, gridColumnSpan: 1, gridColumnStart: 1, msGridRow: 1, msGridRowSpan: 1, msGridColumn: 1, msGridColumnSpan: 1, fontWeight: 1, lineHeight: 1, opacity: 1, order: 1, orphans: 1, tabSize: 1, widows: 1, zIndex: 1, zoom: 1, WebkitLineClamp: 1, fillOpacity: 1, floodOpacity: 1, stopOpacity: 1, strokeDasharray: 1, strokeDashoffset: 1, strokeMiterlimit: 1, strokeOpacity: 1, strokeWidth: 1 }; var u = /^((children|dangerouslySetInnerHTML|key|ref|autoFocus|defaultValue|defaultChecked|innerHTML|suppressContentEditableWarning|suppressHydrationWarning|valueLink|accept|acceptCharset|accessKey|action|allow|allowUserMedia|allowPaymentRequest|allowFullScreen|allowTransparency|alt|async|autoComplete|autoPlay|capture|cellPadding|cellSpacing|challenge|charSet|checked|cite|classID|className|cols|colSpan|content|contentEditable|contextMenu|controls|controlsList|coords|crossOrigin|data|dateTime|decoding|default|defer|dir|disabled|disablePictureInPicture|download|draggable|encType|form|formAction|formEncType|formMethod|formNoValidate|formTarget|frameBorder|headers|height|hidden|high|href|hrefLang|htmlFor|httpEquiv|id|inputMode|integrity|is|keyParams|keyType|kind|label|lang|list|loading|loop|low|marginHeight|marginWidth|max|maxLength|media|mediaGroup|method|min|minLength|multiple|muted|name|nonce|noValidate|open|optimum|pattern|placeholder|playsInline|poster|preload|profile|radioGroup|readOnly|referrerPolicy|rel|required|reversed|role|rows|rowSpan|sandbox|scope|scoped|scrolling|seamless|selected|shape|size|sizes|slot|span|spellCheck|src|srcDoc|srcLang|srcSet|start|step|style|summary|tabIndex|target|title|type|useMap|value|width|wmode|wrap|about|datatype|inlist|prefix|property|resource|typeof|vocab|autoCapitalize|autoCorrect|autoSave|color|inert|itemProp|itemScope|itemType|itemID|itemRef|on|results|security|unselectable|accentHeight|accumulate|additive|alignmentBaseline|allowReorder|alphabetic|amplitude|arabicForm|ascent|attributeName|attributeType|autoReverse|azimuth|baseFrequency|baselineShift|baseProfile|bbox|begin|bias|by|calcMode|capHeight|clip|clipPathUnits|clipPath|clipRule|colorInterpolation|colorInterpolationFilters|colorProfile|colorRendering|contentScriptType|contentStyleType|cursor|cx|cy|d|decelerate|descent|diffuseConstant|direction|display|divisor|dominantBaseline|dur|dx|dy|edgeMode|elevation|enableBackground|end|exponent|externalResourcesRequired|fill|fillOpacity|fillRule|filter|filterRes|filterUnits|floodColor|floodOpacity|focusable|fontFamily|fontSize|fontSizeAdjust|fontStretch|fontStyle|fontVariant|fontWeight|format|from|fr|fx|fy|g1|g2|glyphName|glyphOrientationHorizontal|glyphOrientationVertical|glyphRef|gradientTransform|gradientUnits|hanging|horizAdvX|horizOriginX|ideographic|imageRendering|in|in2|intercept|k|k1|k2|k3|k4|kernelMatrix|kernelUnitLength|kerning|keyPoints|keySplines|keyTimes|lengthAdjust|letterSpacing|lightingColor|limitingConeAngle|local|markerEnd|markerMid|markerStart|markerHeight|markerUnits|markerWidth|mask|maskContentUnits|maskUnits|mathematical|mode|numOctaves|offset|opacity|operator|order|orient|orientation|origin|overflow|overlinePosition|overlineThickness|panose1|paintOrder|pathLength|patternContentUnits|patternTransform|patternUnits|pointerEvents|points|pointsAtX|pointsAtY|pointsAtZ|preserveAlpha|preserveAspectRatio|primitiveUnits|r|radius|refX|refY|renderingIntent|repeatCount|repeatDur|requiredExtensions|requiredFeatures|restart|result|rotate|rx|ry|scale|seed|shapeRendering|slope|spacing|specularConstant|specularExponent|speed|spreadMethod|startOffset|stdDeviation|stemh|stemv|stitchTiles|stopColor|stopOpacity|strikethroughPosition|strikethroughThickness|string|stroke|strokeDasharray|strokeDashoffset|strokeLinecap|strokeLinejoin|strokeMiterlimit|strokeOpacity|strokeWidth|surfaceScale|systemLanguage|tableValues|targetX|targetY|textAnchor|textDecoration|textRendering|textLength|to|transform|u1|u2|underlinePosition|underlineThickness|unicode|unicodeBidi|unicodeRange|unitsPerEm|vAlphabetic|vHanging|vIdeographic|vMathematical|values|vectorEffect|version|vertAdvY|vertOriginX|vertOriginY|viewBox|viewTarget|visibility|widths|wordSpacing|writingMode|x|xHeight|x1|x2|xChannelSelector|xlinkActuate|xlinkArcrole|xlinkHref|xlinkRole|xlinkShow|xlinkTitle|xlinkType|xmlBase|xmlns|xmlnsXlink|xmlLang|xmlSpace|y|y1|y2|yChannelSelector|z|zoomAndPan|for|class|autofocus)|(([Dd][Aa][Tt][Aa]|[Aa][Rr][Ii][Aa]|x)-.*))$/; const d = (f = {}, function (e) { return void 0 === f[e] && (f[e] = (t = e, u.test(t) || 111 === t.charCodeAt(0) && 110 === t.charCodeAt(1) && t.charCodeAt(2) < 91)), f[e]; var t }); var f, p = r(679), h = r.n(p); function g() { return (g = Object.assign || function (e) { for (var t = 1; t < arguments.length; t++) { var r = arguments[t]; for (var n in r) Object.prototype.hasOwnProperty.call(r, n) && (e[n] = r[n]) } return e }).apply(this, arguments) } var m = function (e, t) { for (var r = [e[0]], n = 0, a = t.length; n < a; n += 1)r.push(t[n], e[n + 1]); return r }, y = function (e) { return null !== e && "object" == typeof e && "[object Object]" === (e.toString ? e.toString() : Object.prototype.toString.call(e)) && !(0, o.typeOf)(e) }, b = Object.freeze([]), v = Object.freeze({}); function S(e) { return "function" == typeof e } function w(e) { return e.displayName || e.name || "Component" } function x(e) { return e && "string" == typeof e.styledComponentId } var C = "undefined" != typeof process && (process.env.REACT_APP_SC_ATTR || process.env.SC_ATTR) || "data-styled", k = "undefined" != typeof window && "HTMLElement" in window, A = Boolean("boolean" == typeof SC_DISABLE_SPEEDY ? SC_DISABLE_SPEEDY : "undefined" != typeof process && void 0 !== process.env.REACT_APP_SC_DISABLE_SPEEDY && "" !== process.env.REACT_APP_SC_DISABLE_SPEEDY ? "false" !== process.env.REACT_APP_SC_DISABLE_SPEEDY && process.env.REACT_APP_SC_DISABLE_SPEEDY : "undefined" != typeof process && void 0 !== process.env.SC_DISABLE_SPEEDY && "" !== process.env.SC_DISABLE_SPEEDY && "false" !== process.env.SC_DISABLE_SPEEDY && process.env.SC_DISABLE_SPEEDY); function P(e) { for (var t = arguments.length, r = new Array(t > 1 ? t - 1 : 0), n = 1; n < t; n++)r[n - 1] = arguments[n]; throw new Error("An error occurred. See https://git.io/JUIaE#" + e + " for more information." + (r.length > 0 ? " Args: " + r.join(", ") : "")) } var O = function () { function e(e) { this.groupSizes = new Uint32Array(512), this.length = 512, this.tag = e } var t = e.prototype; return t.indexOfGroup = function (e) { for (var t = 0, r = 0; r < e; r++)t += this.groupSizes[r]; return t }, t.insertRules = function (e, t) { if (e >= this.groupSizes.length) { for (var r = this.groupSizes, n = r.length, a = n; e >= a;)(a <<= 1) < 0 && P(16, "" + e); this.groupSizes = new Uint32Array(a), this.groupSizes.set(r), this.length = a; for (var o = n; o < a; o++)this.groupSizes[o] = 0 } for (var i = this.indexOfGroup(e + 1), s = 0, c = t.length; s < c; s++)this.tag.insertRule(i, t[s]) && (this.groupSizes[e]++, i++) }, t.clearGroup = function (e) { if (e < this.length) { var t = this.groupSizes[e], r = this.indexOfGroup(e), n = r + t; this.groupSizes[e] = 0; for (var a = r; a < n; a++)this.tag.deleteRule(r) } }, t.getGroup = function (e) { var t = ""; if (e >= this.length || 0 === this.groupSizes[e]) return t; for (var r = this.groupSizes[e], n = this.indexOfGroup(e), a = n + r, o = n; o < a; o++)t += this.tag.getRule(o) + "/*!sc*/\n"; return t }, e }(), I = new Map, R = new Map, E = 1, $ = function (e) { if (I.has(e)) return I.get(e); for (; R.has(E);)E++; var t = E++; return I.set(e, t), R.set(t, e), t }, T = function (e) { return R.get(e) }, _ = function (e, t) { I.set(e, t), R.set(t, e) }, j = "style[" + C + '][data-styled-version="5.2.1"]', M = new RegExp("^" + C + '\\.g(\\d+)\\[id="([\\w\\d-]+)"\\].*?"([^"]*)'), B = function (e, t, r) { for (var n, a = r.split(","), o = 0, i = a.length; o < i; o++)(n = a[o]) && e.registerName(t, n) }, N = function (e, t) { for (var r = t.innerHTML.split("/*!sc*/\n"), n = [], a = 0, o = r.length; a < o; a++) { var i = r[a].trim(); if (i) { var s = i.match(M); if (s) { var c = 0 | parseInt(s[1], 10), l = s[2]; 0 !== c && (_(l, c), B(e, l, s[3]), e.getTag().insertRules(c, n)), n.length = 0 } else n.push(i) } } }, z = function () { return r.nc }, L = function (e) { var t = document.head, r = e || t, n = document.createElement("style"), a = function (e) { for (var t = e.childNodes, r = t.length; r >= 0; r--) { var n = t[r]; if (n && 1 === n.nodeType && n.hasAttribute(C)) return n } }(r), o = void 0 !== a ? a.nextSibling : null; n.setAttribute(C, "active"), n.setAttribute("data-styled-version", "5.2.1"); var i = z(); return i && n.setAttribute("nonce", i), r.insertBefore(n, o), n }, F = function () { function e(e) { var t = this.element = L(e); t.appendChild(document.createTextNode("")), this.sheet = function (e) { if (e.sheet) return e.sheet; for (var t = document.styleSheets, r = 0, n = t.length; r < n; r++) { var a = t[r]; if (a.ownerNode === e) return a } P(17) }(t), this.length = 0 } var t = e.prototype; return t.insertRule = function (e, t) { try { return this.sheet.insertRule(t, e), this.length++, !0 } catch (e) { return !1 } }, t.deleteRule = function (e) { this.sheet.deleteRule(e), this.length-- }, t.getRule = function (e) { var t = this.sheet.cssRules[e]; return void 0 !== t && "string" == typeof t.cssText ? t.cssText : "" }, e }(), D = function () { function e(e) { var t = this.element = L(e); this.nodes = t.childNodes, this.length = 0 } var t = e.prototype; return t.insertRule = function (e, t) { if (e <= this.length && e >= 0) { var r = document.createTextNode(t), n = this.nodes[e]; return this.element.insertBefore(r, n || null), this.length++, !0 } return !1 }, t.deleteRule = function (e) { this.element.removeChild(this.nodes[e]), this.length-- }, t.getRule = function (e) { return e < this.length ? this.nodes[e].textContent : "" }, e }(), H = function () { function e(e) { this.rules = [], this.length = 0 } var t = e.prototype; return t.insertRule = function (e, t) { return e <= this.length && (this.rules.splice(e, 0, t), this.length++, !0) }, t.deleteRule = function (e) { this.rules.splice(e, 1), this.length-- }, t.getRule = function (e) { return e < this.length ? this.rules[e] : "" }, e }(), G = k, W = { isServer: !k, useCSSOMInjection: !A }, Y = function () { function e(e, t, r) { void 0 === e && (e = v), void 0 === t && (t = {}), this.options = g({}, W, {}, e), this.gs = t, this.names = new Map(r), !this.options.isServer && k && G && (G = !1, function (e) { for (var t = document.querySelectorAll(j), r = 0, n = t.length; r < n; r++) { var a = t[r]; a && "active" !== a.getAttribute(C) && (N(e, a), a.parentNode && a.parentNode.removeChild(a)) } }(this)) } e.registerId = function (e) { return $(e) }; var t = e.prototype; return t.reconstructWithOptions = function (t, r) { return void 0 === r && (r = !0), new e(g({}, this.options, {}, t), this.gs, r && this.names || void 0) }, t.allocateGSInstance = function (e) { return this.gs[e] = (this.gs[e] || 0) + 1 }, t.getTag = function () { return this.tag || (this.tag = (r = (t = this.options).isServer, n = t.useCSSOMInjection, a = t.target, e = r ? new H(a) : n ? new F(a) : new D(a), new O(e))); var e, t, r, n, a }, t.hasNameForId = function (e, t) { return this.names.has(e) && this.names.get(e).has(t) }, t.registerName = function (e, t) { if ($(e), this.names.has(e)) this.names.get(e).add(t); else { var r = new Set; r.add(t), this.names.set(e, r) } }, t.insertRules = function (e, t, r) { this.registerName(e, t), this.getTag().insertRules($(e), r) }, t.clearNames = function (e) { this.names.has(e) && this.names.get(e).clear() }, t.clearRules = function (e) { this.getTag().clearGroup($(e)), this.clearNames(e) }, t.clearTag = function () { this.tag = void 0 }, t.toString = function () { return function (e) { for (var t = e.getTag(), r = t.length, n = "", a = 0; a < r; a++) { var o = T(a); if (void 0 !== o) { var i = e.names.get(o), s = t.getGroup(a); if (void 0 !== i && 0 !== s.length) { var c = C + ".g" + a + '[id="' + o + '"]', l = ""; void 0 !== i && i.forEach((function (e) { e.length > 0 && (l += e + ",") })), n += "" + s + c + '{content:"' + l + '"}/*!sc*/\n' } } } return n }(this) }, e }(), U = /(a)(d)/gi, q = function (e) { return String.fromCharCode(e + (e > 25 ? 39 : 97)) }; function V(e) { var t, r = ""; for (t = Math.abs(e); t > 52; t = t / 52 | 0)r = q(t % 52) + r; return (q(t % 52) + r).replace(U, "$1-$2") } var X = function (e, t) { for (var r = t.length; r;)e = 33 * e ^ t.charCodeAt(--r); return e }, Z = function (e) { return X(5381, e) }; function J(e) { for (var t = 0; t < e.length; t += 1) { var r = e[t]; if (S(r) && !x(r)) return !1 } return !0 } var K = Z("5.2.1"), Q = function () { function e(e, t, r) { this.rules = e, this.staticRulesId = "", this.isStatic = (void 0 === r || r.isStatic) && J(e), this.componentId = t, this.baseHash = X(K, t), this.baseStyle = r, Y.registerId(t) } return e.prototype.generateAndInjectStyles = function (e, t, r) { var n = this.componentId, a = []; if (this.baseStyle && a.push(this.baseStyle.generateAndInjectStyles(e, t, r)), this.isStatic && !r.hash) if (this.staticRulesId && t.hasNameForId(n, this.staticRulesId)) a.push(this.staticRulesId); else { var o = me(this.rules, e, t, r).join(""), i = V(X(this.baseHash, o.length) >>> 0); if (!t.hasNameForId(n, i)) { var s = r(o, "." + i, void 0, n); t.insertRules(n, i, s) } a.push(i), this.staticRulesId = i } else { for (var c = this.rules.length, l = X(this.baseHash, r.hash), u = "", d = 0; d < c; d++) { var f = this.rules[d]; if ("string" == typeof f) u += f; else if (f) { var p = me(f, e, t, r), h = Array.isArray(p) ? p.join("") : p; l = X(l, h + d), u += h } } if (u) { var g = V(l >>> 0); if (!t.hasNameForId(n, g)) { var m = r(u, "." + g, void 0, n); t.insertRules(n, g, m) } a.push(g) } } return a.join(" ") }, e }(), ee = /^\s*\/\/.*$/gm, te = [":", "[", ".", "#"]; function re(e) { var t, r, n, a, o = void 0 === e ? v : e, i = o.options, s = void 0 === i ? v : i, l = o.plugins, u = void 0 === l ? b : l, d = new c(s), f = [], p = function (e) { function t(t) { if (t) try { e(t + "}") } catch (e) { } } return function (r, n, a, o, i, s, c, l, u, d) { switch (r) { case 1: if (0 === u && 64 === n.charCodeAt(0)) return e(n + ";"), ""; break; case 2: if (0 === l) return n + "/*|*/"; break; case 3: switch (l) { case 102: case 112: return e(a[0] + n), ""; default: return n + (0 === d ? "/*|*/" : "") }case -2: n.split("/*|*/}").forEach(t) } } }((function (e) { f.push(e) })), h = function (e, n, o) { return 0 === n && te.includes(o[r.length]) || o.match(a) ? e : "." + t }; function g(e, o, i, s) { void 0 === s && (s = "&"); var c = e.replace(ee, ""), l = o && i ? i + " " + o + " { " + c + " }" : c; return t = s, r = o, n = new RegExp("\\" + r + "\\b", "g"), a = new RegExp("(\\" + r + "\\b){2,}"), d(i || !o ? "" : o, l) } return d.use([].concat(u, [function (e, t, a) { 2 === e && a.length && a[0].lastIndexOf(r) > 0 && (a[0] = a[0].replace(n, h)) }, p, function (e) { if (-2 === e) { var t = f; return f = [], t } }])), g.hash = u.length ? u.reduce((function (e, t) { return t.name || P(15), X(e, t.name) }), 5381).toString() : "", g } var ne = a().createContext(), ae = (ne.Consumer, a().createContext()), oe = (ae.Consumer, new Y), ie = re(); function se() { return (0, n.useContext)(ne) || oe } function ce(e) { var t = (0, n.useState)(e.stylisPlugins), r = t[0], o = t[1], i = se(), c = (0, n.useMemo)((function () { var t = i; return e.sheet ? t = e.sheet : e.target && (t = t.reconstructWithOptions({ target: e.target }, !1)), e.disableCSSOMInjection && (t = t.reconstructWithOptions({ useCSSOMInjection: !1 })), t }), [e.disableCSSOMInjection, e.sheet, e.target]), l = (0, n.useMemo)((function () { return re({ options: { prefix: !e.disableVendorPrefixes }, plugins: r }) }), [e.disableVendorPrefixes, r]); return (0, n.useEffect)((function () { s()(r, e.stylisPlugins) || o(e.stylisPlugins) }), [e.stylisPlugins]), a().createElement(ne.Provider, { value: c }, a().createElement(ae.Provider, { value: l }, e.children)) } var le = function () { function e(e, t) { var r = this; this.inject = function (e, t) { void 0 === t && (t = ie); var n = r.name + t.hash; e.hasNameForId(r.id, n) || e.insertRules(r.id, n, t(r.rules, n, "@keyframes")) }, this.toString = function () { return P(12, String(r.name)) }, this.name = e, this.id = "sc-keyframes-" + e, this.rules = t } return e.prototype.getName = function (e) { return void 0 === e && (e = ie), this.name + e.hash }, e }(), ue = /([A-Z])/, de = /([A-Z])/g, fe = /^ms-/, pe = function (e) { return "-" + e.toLowerCase() }; function he(e) { return ue.test(e) ? e.replace(de, pe).replace(fe, "-ms-") : e } var ge = function (e) { return null == e || !1 === e || "" === e }; function me(e, t, r, n) { if (Array.isArray(e)) { for (var a, o = [], i = 0, s = e.length; i < s; i += 1)"" !== (a = me(e[i], t, r, n)) && (Array.isArray(a) ? o.push.apply(o, a) : o.push(a)); return o } return ge(e) ? "" : x(e) ? "." + e.styledComponentId : S(e) ? "function" != typeof (c = e) || c.prototype && c.prototype.isReactComponent || !t ? e : me(e(t), t, r, n) : e instanceof le ? r ? (e.inject(r, n), e.getName(n)) : e : y(e) ? function e(t, r) { var n, a, o = []; for (var i in t) t.hasOwnProperty(i) && !ge(t[i]) && (y(t[i]) ? o.push.apply(o, e(t[i], i)) : S(t[i]) ? o.push(he(i) + ":", t[i], ";") : o.push(he(i) + ": " + (n = i, (null == (a = t[i]) || "boolean" == typeof a || "" === a ? "" : "number" != typeof a || 0 === a || n in l ? String(a).trim() : a + "px") + ";"))); return r ? [r + " {"].concat(o, ["}"]) : o }(e) : e.toString(); var c } function ye(e) { for (var t = arguments.length, r = new Array(t > 1 ? t - 1 : 0), n = 1; n < t; n++)r[n - 1] = arguments[n]; return S(e) || y(e) ? me(m(b, [e].concat(r))) : 0 === r.length && 1 === e.length && "string" == typeof e[0] ? e : me(m(e, r)) } new Set; var be = /[!"#$%&'()*+,./:;<=>?@[\\\]^`{|}~-]+/g, ve = /(^-|-$)/g; function Se(e) { return e.replace(be, "-").replace(ve, "") } function we(e) { return "string" == typeof e && !0 } var xe = function (e) { return "function" == typeof e || "object" == typeof e && null !== e && !Array.isArray(e) }, Ce = function (e) { return "__proto__" !== e && "constructor" !== e && "prototype" !== e }; function ke(e, t, r) { var n = e[r]; xe(t) && xe(n) ? Ae(n, t) : e[r] = t } function Ae(e) { for (var t = arguments.length, r = new Array(t > 1 ? t - 1 : 0), n = 1; n < t; n++)r[n - 1] = arguments[n]; for (var a = 0, o = r; a < o.length; a++) { var i = o[a]; if (xe(i)) for (var s in i) Ce(s) && ke(e, i[s], s) } return e } var Pe = a().createContext(); Pe.Consumer; var Oe = {}; function Ie(e, t, r) { var o = x(e), i = !we(e), s = t.attrs, c = void 0 === s ? b : s, l = t.componentId, u = void 0 === l ? function (e, t) { var r = "string" != typeof e ? "sc" : Se(e); Oe[r] = (Oe[r] || 0) + 1; var n = r + "-" + function (e) { return V(Z(e) >>> 0) }("5.2.1" + r + Oe[r]); return t ? t + "-" + n : n }(t.displayName, t.parentComponentId) : l, f = t.displayName, p = void 0 === f ? function (e) { return we(e) ? "styled." + e : "Styled(" + w(e) + ")" }(e) : f, m = t.displayName && t.componentId ? Se(t.displayName) + "-" + t.componentId : t.componentId || u, y = o && e.attrs ? Array.prototype.concat(e.attrs, c).filter(Boolean) : c, C = t.shouldForwardProp; o && e.shouldForwardProp && (C = t.shouldForwardProp ? function (r, n) { return e.shouldForwardProp(r, n) && t.shouldForwardProp(r, n) } : e.shouldForwardProp); var k, A = new Q(r, m, o ? e.componentStyle : void 0), P = A.isStatic && 0 === c.length, O = function (e, t) { return function (e, t, r, a) { var o = e.attrs, i = e.componentStyle, s = e.defaultProps, c = e.foldedComponentIds, l = e.shouldForwardProp, u = e.styledComponentId, f = e.target, p = function (e, t, r) { void 0 === e && (e = v); var n = g({}, t, { theme: e }), a = {}; return r.forEach((function (e) { var t, r, o, i = e; for (t in S(i) && (i = i(n)), i) n[t] = a[t] = "className" === t ? (r = a[t], o = i[t], r && o ? r + " " + o : r || o) : i[t] })), [n, a] }(function (e, t, r) { return void 0 === r && (r = v), e.theme !== r.theme && e.theme || t || r.theme }(t, (0, n.useContext)(Pe), s) || v, t, o), h = p[0], m = p[1], y = function (e, t, r, a) { var o = se(), i = (0, n.useContext)(ae) || ie; return t ? e.generateAndInjectStyles(v, o, i) : e.generateAndInjectStyles(r, o, i) }(i, a, h), b = r, w = m.$as || t.$as || m.as || t.as || f, x = we(w), C = m !== t ? g({}, t, {}, m) : t, k = {}; for (var A in C) "$" !== A[0] && "as" !== A && ("forwardedAs" === A ? k.as = C[A] : (l ? l(A, d) : !x || d(A)) && (k[A] = C[A])); return t.style && m.style !== t.style && (k.style = g({}, t.style, {}, m.style)), k.className = Array.prototype.concat(c, u, y !== u ? y : null, t.className, m.className).filter(Boolean).join(" "), k.ref = b, (0, n.createElement)(w, k) }(k, e, t, P) }; return O.displayName = p, (k = a().forwardRef(O)).attrs = y, k.componentStyle = A, k.displayName = p, k.shouldForwardProp = C, k.foldedComponentIds = o ? Array.prototype.concat(e.foldedComponentIds, e.styledComponentId) : b, k.styledComponentId = m, k.target = o ? e.target : e, k.withComponent = function (e) { var n = t.componentId, a = function (e, t) { if (null == e) return {}; var r, n, a = {}, o = Object.keys(e); for (n = 0; n < o.length; n++)r = o[n], t.indexOf(r) >= 0 || (a[r] = e[r]); return a }(t, ["componentId"]), o = n && n + "-" + (we(e) ? e : Se(w(e))); return Ie(e, g({}, a, { attrs: y, componentId: o }), r) }, Object.defineProperty(k, "defaultProps", { get: function () { return this._foldedDefaultProps }, set: function (t) { this._foldedDefaultProps = o ? Ae({}, e.defaultProps, t) : t } }), k.toString = function () { return "." + k.styledComponentId }, i && h()(k, e, { attrs: !0, componentStyle: !0, displayName: !0, foldedComponentIds: !0, shouldForwardProp: !0, styledComponentId: !0, target: !0, withComponent: !0 }), k } var Re, Ee = function (e) { return function e(t, r, n) { if (void 0 === n && (n = v), !(0, o.isValidElementType)(r)) return P(1, String(r)); var a = function () { return t(r, n, ye.apply(void 0, arguments)) }; return a.withConfig = function (a) { return e(t, r, g({}, n, {}, a)) }, a.attrs = function (a) { return e(t, r, g({}, n, { attrs: Array.prototype.concat(n.attrs, a).filter(Boolean) })) }, a }(Ie, e) };["a", "abbr", "address", "area", "article", "aside", "audio", "b", "base", "bdi", "bdo", "big", "blockquote", "body", "br", "button", "canvas", "caption", "cite", "code", "col", "colgroup", "data", "datalist", "dd", "del", "details", "dfn", "dialog", "div", "dl", "dt", "em", "embed", "fieldset", "figcaption", "figure", "footer", "form", "h1", "h2", "h3", "h4", "h5", "h6", "head", "header", "hgroup", "hr", "html", "i", "iframe", "img", "input", "ins", "kbd", "keygen", "label", "legend", "li", "link", "main", "map", "mark", "marquee", "menu", "menuitem", "meta", "meter", "nav", "noscript", "object", "ol", "optgroup", "option", "output", "p", "param", "picture", "pre", "progress", "q", "rp", "rt", "ruby", "s", "samp", "script", "section", "select", "small", "source", "span", "strong", "style", "sub", "summary", "sup", "table", "tbody", "td", "textarea", "tfoot", "th", "thead", "time", "title", "tr", "track", "u", "ul", "var", "video", "wbr", "circle", "clipPath", "defs", "ellipse", "foreignObject", "g", "image", "line", "linearGradient", "marker", "mask", "path", "pattern", "polygon", "polyline", "radialGradient", "rect", "stop", "svg", "text", "tspan"].forEach((function (e) { Ee[e] = Ee(e) })), (Re = function (e, t) { this.rules = e, this.componentId = t, this.isStatic = J(e), Y.registerId(this.componentId + 1) }.prototype).createStyles = function (e, t, r, n) { var a = n(me(this.rules, t, r, n).join(""), ""), o = this.componentId + e; r.insertRules(o, o, a) }, Re.removeStyles = function (e, t) { t.clearRules(this.componentId + e) }, Re.renderStyles = function (e, t, r, n) { e > 2 && Y.registerId(this.componentId + e), this.removeStyles(e, r), this.createStyles(e, t, r, n) }, function () { var e = function () { var e = this; this._emitSheetCSS = function () { var t = e.instance.toString(), r = z(); return "<style " + [r && 'nonce="' + r + '"', C + '="true"', 'data-styled-version="5.2.1"'].filter(Boolean).join(" ") + ">" + t + "</style>" }, this.getStyleTags = function () { return e.sealed ? P(2) : e._emitSheetCSS() }, this.getStyleElement = function () { var t; if (e.sealed) return P(2); var r = ((t = {})[C] = "", t["data-styled-version"] = "5.2.1", t.dangerouslySetInnerHTML = { __html: e.instance.toString() }, t), n = z(); return n && (r.nonce = n), [a().createElement("style", g({}, r, { key: "sc-0-0" }))] }, this.seal = function () { e.sealed = !0 }, this.instance = new Y({ isServer: !0 }), this.sealed = !1 }.prototype; e.collectStyles = function (e) { return this.sealed ? P(2) : a().createElement(ce, { sheet: this.instance }, e) }, e.interleaveWithNodeStream = function (e) { return P(3) } }(); const $e = Ee, Te = $e.div.withConfig({ displayName: "Container" })(["width:", ";min-height:", ";height:auto;border-radius:", ";border:1px solid ", ";background-color:", ";backdrop-filter:", ";z-index:100;", " margin-bottom:30px;display:grid;grid-template-columns:repeat(2,80% 20%);"], (function (e) { return e.cardWidth && e.cardWidth > 0 ? "".concat(e.cardWidth, "px") : "300px" }), (function (e) { return e.cardHeight && e.cardHeight > 0 ? "".concat(e.cardHeight, "px") : "150px" }), (function (e) { return e.cardRadius && e.cardRadius > 0 ? "".concat(e.cardRadius, "px") : "15px" }), (function (e) { return e.cardColorBorder && "" !== e.cardColorBorder ? e.cardColorBorder : "#f3f2f145" }), (function (e) { return e.cardBackground && "" !== e.cardBackground.trim() ? e.cardBackground : "#f3f2f108" }), (function (e) { return e.cardGlass && e.cardGlass > 0 ? "blur(".concat(e.cardGlass, "px)") : "blur(5px)" }), (function (e) { return e.cardPaddingAll && e.cardPaddingAll >= 0 ? ye(["padding:", "px;"], e.cardPaddingAll) : e.cardPaddingLeft && e.cardPaddingRight && e.cardPaddingTop && e.cardPaddingBottom ? ye(["padding:", "px ", "px ", "px ", "px"], e.cardPaddingTop, e.cardPaddingLeft, e.cardPaddingBottom, e.cardPaddingRight) : ye(["padding:10px 20px 10px 20px;"]) })); var _e = $e.div.withConfig({ displayName: "styles__Section" })(["width:100%;min-height:100%;height:auto;& + &{position:relative;}display:flex;flex-flow:column;"]), je = $e.div.withConfig({ displayName: "styles__BoxContainer" })(["width:100%;height:auto;", " display:flex;flex-flow:column;"], (function (e) { return e.boxChatOnMe && !0 === e.boxChatOnMe && ye(['&::after{content:"";position:absolute;background:none;right:0px;bottom:0px;}']) })), Me = ($e.div.withConfig({ displayName: "styles__BoxHeader" })(["width:100%;height:auto;span{font-family:\"'Lato Hairline', Lato\",sans-serif;font-weight:700;font-size:0.9em;color:#FFF;letter-spacing:0.7px;}"]), $e.div.withConfig({ displayName: "styles__BoxContent" })(["width:100%;flex:1;span{font-family:\"'Lato Hairline', Lato\",sans-serif;font-weight:lighter;font-size:0.6em;color:#FFF;letter-spacing:0.7px;}"])); $e.button.withConfig({ displayName: "styles__Button" })(["width:100%;height:25px;background-color:", ";cursor:pointer;position:absolute;bottom:0px;left:10px;font-family:\"'Lato Hairline', Lato\",sans-serif;font-weight:600;font-size:0.5em;color:", ";letter-spacing:0.7px;display:flex;flex-flow:row;justify-content:center;align-items:center;"], (function (e) { return e.boxMessageButtonColor && "" !== e.boxMessageButtonColor ? e.boxMessageButtonColor : "#4F3AA7" }), (function (e) { return e.boxMessageButtonTextColor && "" !== e.boxMessageButtonTextColor ? e.boxMessageButtonTextColor : "#FFF" })); const Be = function (e) { e.boxChatTitle; var t = e.boxChatContent, r = e.boxChatOnMe, n = e.cardWidth, o = e.cardHeight, i = e.cardRadius, s = e.cardBackground, c = e.cardGlass, l = e.cardColorBorder, u = e.cardPaddingAll, d = e.cardPaddingLeft, f = e.cardPaddingRight, p = e.cardPaddingTop, h = e.cardPaddingBottom; return e.children, function (e, t) { if (null == e) return {}; var r, n, a = function (e, t) { if (null == e) return {}; var r, n, a = {}, o = Object.keys(e); for (n = 0; n < o.length; n++)r = o[n], t.indexOf(r) >= 0 || (a[r] = e[r]); return a }(e, t); if (Object.getOwnPropertySymbols) { var o = Object.getOwnPropertySymbols(e); for (n = 0; n < o.length; n++)r = o[n], t.indexOf(r) >= 0 || Object.prototype.propertyIsEnumerable.call(e, r) && (a[r] = e[r]) } }(e, ["boxChatTitle", "boxChatContent", "boxChatOnMe", "cardWidth", "cardHeight", "cardRadius", "cardBackground", "cardGlass", "cardColorBorder", "cardPaddingAll", "cardPaddingLeft", "cardPaddingRight", "cardPaddingTop", "cardPaddingBottom", "children"]), a().createElement(Te, { cardWidth: n, cardHeight: o, cardRadius: i, cardGlass: c, cardBackground: s, cardColorBorder: l, cardPaddingAll: u, cardPaddingLeft: d, cardPaddingRight: f, cardPaddingTop: p, cardPaddingBottom: h }, a().createElement(_e, null, a().createElement(je, { boxChatOnMe: r }, a().createElement(Me, null, a().createElement("span", null, t)))), a().createElement(_e, null)) } }, 679: (e, t, r) => { "use strict"; var n = r(296), a = { childContextTypes: !0, contextType: !0, contextTypes: !0, defaultProps: !0, displayName: !0, getDefaultProps: !0, getDerivedStateFromError: !0, getDerivedStateFromProps: !0, mixins: !0, propTypes: !0, type: !0 }, o = { name: !0, length: !0, prototype: !0, caller: !0, callee: !0, arguments: !0, arity: !0 }, i = { $$typeof: !0, compare: !0, defaultProps: !0, displayName: !0, propTypes: !0, type: !0 }, s = {}; function c(e) { return n.isMemo(e) ? i : s[e.$$typeof] || a } s[n.ForwardRef] = { $$typeof: !0, render: !0, defaultProps: !0, displayName: !0, propTypes: !0 }, s[n.Memo] = i; var l = Object.defineProperty, u = Object.getOwnPropertyNames, d = Object.getOwnPropertySymbols, f = Object.getOwnPropertyDescriptor, p = Object.getPrototypeOf, h = Object.prototype; e.exports = function e(t, r, n) { if ("string" != typeof r) { if (h) { var a = p(r); a && a !== h && e(t, a, n) } var i = u(r); d && (i = i.concat(d(r))); for (var s = c(t), g = c(r), m = 0; m < i.length; ++m) { var y = i[m]; if (!(o[y] || n && n[y] || g && g[y] || s && s[y])) { var b = f(r, y); try { l(t, y, b) } catch (e) { } } } } return t } }, 103: (e, t) => { "use strict"; var r = "function" == typeof Symbol && Symbol.for, n = r ? Symbol.for("react.element") : 60103, a = r ? Symbol.for("react.portal") : 60106, o = r ? Symbol.for("react.fragment") : 60107, i = r ? Symbol.for("react.strict_mode") : 60108, s = r ? Symbol.for("react.profiler") : 60114, c = r ? Symbol.for("react.provider") : 60109, l = r ? Symbol.for("react.context") : 60110, u = r ? Symbol.for("react.async_mode") : 60111, d = r ? Symbol.for("react.concurrent_mode") : 60111, f = r ? Symbol.for("react.forward_ref") : 60112, p = r ? Symbol.for("react.suspense") : 60113, h = r ? Symbol.for("react.suspense_list") : 60120, g = r ? Symbol.for("react.memo") : 60115, m = r ? Symbol.for("react.lazy") : 60116, y = r ? Symbol.for("react.block") : 60121, b = r ? Symbol.for("react.fundamental") : 60117, v = r ? Symbol.for("react.responder") : 60118, S = r ? Symbol.for("react.scope") : 60119; function w(e) { if ("object" == typeof e && null !== e) { var t = e.$$typeof; switch (t) { case n: switch (e = e.type) { case u: case d: case o: case s: case i: case p: return e; default: switch (e = e && e.$$typeof) { case l: case f: case m: case g: case c: return e; default: return t } }case a: return t } } } function x(e) { return w(e) === d } t.AsyncMode = u, t.ConcurrentMode = d, t.ContextConsumer = l, t.ContextProvider = c, t.Element = n, t.ForwardRef = f, t.Fragment = o, t.Lazy = m, t.Memo = g, t.Portal = a, t.Profiler = s, t.StrictMode = i, t.Suspense = p, t.isAsyncMode = function (e) { return x(e) || w(e) === u }, t.isConcurrentMode = x, t.isContextConsumer = function (e) { return w(e) === l }, t.isContextProvider = function (e) { return w(e) === c }, t.isElement = function (e) { return "object" == typeof e && null !== e && e.$$typeof === n }, t.isForwardRef = function (e) { return w(e) === f }, t.isFragment = function (e) { return w(e) === o }, t.isLazy = function (e) { return w(e) === m }, t.isMemo = function (e) { return w(e) === g }, t.isPortal = function (e) { return w(e) === a }, t.isProfiler = function (e) { return w(e) === s }, t.isStrictMode = function (e) { return w(e) === i }, t.isSuspense = function (e) { return w(e) === p }, t.isValidElementType = function (e) { return "string" == typeof e || "function" == typeof e || e === o || e === d || e === s || e === i || e === p || e === h || "object" == typeof e && null !== e && (e.$$typeof === m || e.$$typeof === g || e.$$typeof === c || e.$$typeof === l || e.$$typeof === f || e.$$typeof === b || e.$$typeof === v || e.$$typeof === S || e.$$typeof === y) }, t.typeOf = w }, 296: (e, t, r) => { "use strict"; e.exports = r(103) }, 921: (e, t) => { "use strict"; var r = 60103, n = 60106, a = 60107, o = 60108, i = 60114, s = 60109, c = 60110, l = 60112, u = 60113, d = 60120, f = 60115, p = 60116, h = 60121, g = 60122, m = 60117, y = 60129, b = 60131; if ("function" == typeof Symbol && Symbol.for) { var v = Symbol.for; r = v("react.element"), n = v("react.portal"), a = v("react.fragment"), o = v("react.strict_mode"), i = v("react.profiler"), s = v("react.provider"), c = v("react.context"), l = v("react.forward_ref"), u = v("react.suspense"), d = v("react.suspense_list"), f = v("react.memo"), p = v("react.lazy"), h = v("react.block"), g = v("react.server.block"), m = v("react.fundamental"), y = v("react.debug_trace_mode"), b = v("react.legacy_hidden") } t.isValidElementType = function (e) { return "string" == typeof e || "function" == typeof e || e === a || e === i || e === y || e === o || e === u || e === d || e === b || "object" == typeof e && null !== e && (e.$$typeof === p || e.$$typeof === f || e.$$typeof === s || e.$$typeof === c || e.$$typeof === l || e.$$typeof === m || e.$$typeof === h || e[0] === g) }, t.typeOf = function (e) { if ("object" == typeof e && null !== e) { var t = e.$$typeof; switch (t) { case r: switch (e = e.type) { case a: case i: case o: case u: case d: return e; default: switch (e = e && e.$$typeof) { case c: case l: case p: case f: case s: return e; default: return t } }case n: return t } } } }, 864: (e, t, r) => { "use strict"; e.exports = r(921) }, 774: e => { e.exports = function (e, t, r, n) { var a = r ? r.call(n, e, t) : void 0; if (void 0 !== a) return !!a; if (e === t) return !0; if ("object" != typeof e || !e || "object" != typeof t || !t) return !1; var o = Object.keys(e), i = Object.keys(t); if (o.length !== i.length) return !1; for (var s = Object.prototype.hasOwnProperty.bind(t), c = 0; c < o.length; c++) { var l = o[c]; if (!s(l)) return !1; var u = e[l], d = t[l]; if (!1 === (a = r ? r.call(n, u, d, l) : void 0) || void 0 === a && u !== d) return !1 } return !0 } } }, t = {}; function r(n) { if (t[n]) return t[n].exports; var a = t[n] = { exports: {} }; return e[n](a, a.exports, r), a.exports } return r.n = e => { var t = e && e.__esModule ? () => e.default : () => e; return r.d(t, { a: t }), t }, r.d = (e, t) => { for (var n in t) r.o(t, n) && !r.o(e, n) && Object.defineProperty(e, n, { enumerable: !0, get: t[n] }) }, r.o = (e, t) => Object.prototype.hasOwnProperty.call(e, t), r.r = e => { "undefined" != typeof Symbol && Symbol.toStringTag && Object.defineProperty(e, Symbol.toStringTag, { value: "Module" }), Object.defineProperty(e, "__esModule", { value: !0 }) }, r(646) })();
