--- packages/GlassComponent/src/index.ts
@@ -1,12 +1,0 @@
-import React from "react"
-
-// Components
-import creditCard from "../CreditCard"
-import inputGlass from "../InputGlass"
-import boxChatGlass from "../BoxChatGlass"
-import boxMessageGlass from "../BoxMessageGlass"
-
-export const InputGlass: React.FC = inputGlass;
-export const CreditCardGlass: React.FC = creditCard;
-export const BoxChatGlass = boxChatGlass;
-export const BoxMessageGlass: React.FC = boxMessageGlass;
